requirements.txt
# BITTEN Core System Dependencies
flask>=2.3.0
requests>=2.31.0
python-telegram-bot>=20.5
python-dotenv>=1.0.0
pyyaml>=6.0
pytz>=2023.3
sqlalchemy>=2.0.0
alembic>=1.13.0
cryptography>=41.0.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.9.0